requirements.txt
requests==2.32.3
PyJWT==2.9.0
python-dotenv==1.0.1
discord.py==2.4.0